fixed video duration

+++ features/media/mediapicker/src/main/java/com/tokopedia/media/picker/ui/fragment/camera/component/CameraPreviewComponent.kt
@@ -58,7 +58,7 @@ class CameraPreviewComponent(
 
         cameraView.takeVideoSnapshot(
             FileGenerator.createFileVideoRecorder(),
-            param.maxVideoDuration()
+            param.maxVideoDuration().toInt()
         )
     }
